+++ src/main/java/com/hyperfactions/gui/faction/ChunkMapAsset.java
@@ -3,7 +3,7 @@ package com.hyperfactions.gui.faction;
 import com.hyperfactions.util.Logger;
 import com.hypixel.hytale.common.util.ArrayUtil;
 import com.hypixel.hytale.math.util.ChunkUtil;
-import com.hypixel.hytale.protocol.Packet;
+import com.hypixel.hytale.protocol.ToClientPacket;
 import com.hypixel.hytale.protocol.packets.setup.AssetFinalize;
 import com.hypixel.hytale.protocol.packets.setup.AssetInitialize;
 import com.hypixel.hytale.protocol.packets.setup.AssetPart;
@@ -154,7 +154,7 @@ public class ChunkMapAsset extends CommonAsset {
     public static void sendToPlayer(PacketHandler handler, CommonAsset asset) {
         byte[] allBytes = asset.getBlob().join();
         byte[][] parts = ArrayUtil.split(allBytes, 2621440);
-        Packet[] packets = new Packet[2 + parts.length];
+        ToClientPacket[] packets = new ToClientPacket[2 + parts.length];
         packets[0] = new AssetInitialize(asset.toPacket(), allBytes.length);
 
         for (int partIndex = 0; partIndex < parts.length; ++partIndex) {